+++ src/generic_front/timeline.rs
@@ -114,12 +114,12 @@ impl Timeline {
         let mut lines = iced::widget::canvas::Frame::new(renderer, bounds.size());
 
         let mut beat = Position::from_interleaved_samples(
-            self.position.x as u32,
+            self.position.x as usize,
             &self.arrangement.read().unwrap().meter,
         );
         let mut end_beat = beat
             + Position::from_interleaved_samples(
-                (bounds.width * self.scale.x.exp2()) as u32,
+                (bounds.width * self.scale.x.exp2()) as usize,
                 &self.arrangement.read().unwrap().meter,
             );
         if beat.sub_quarter_note != 0 {
